Rearrange <MediaPage>; nix <Header> from <Home>
Now, <MediaPage> invokes <MediaPageQueryRenderer>, with the new Relay
Modern query [#6211].

<ProjectHeader>, <TeamPublicHeader> and <Header> were a knot of
confusion. Really, the only page that showed a header was <MediaPage>.
So now we can clean up <Home>.

(In the future, for #6211, we'll nix the DrawerNavigation from <Home>,
too.)

[REBASE NOTES] Rebasing atop #8237. This will be partially broken until
later in the rebase. The ProjectMediaProject mutations rely on arguments
that aren't part of the API yet.